feat(seo): split suppliers directory page, extract client SupplierLogo, add dynamic metadata and catalog schemas
